+++ lib/cli.dart
@@ -1,8 +1,8 @@
 library tota.cli;
 
+export 'src/commands/build.dart';
 export 'src/commands/init.dart';
 export 'src/commands/new.dart';
-export 'src/commands/build.dart';
 
 final String title = 'Tota';
 final String docsUrl = 'https://TODO/docs';