+++ lib/src/commands/init.dart
@@ -1,6 +1,7 @@
 import 'package:args/command_runner.dart';
 import 'package:cli_util/cli_logging.dart';
 import 'package:path/path.dart' as p;
+
 import '../../tota.dart' as tota;
 
 class InitCommand extends Command {